+++ spec/avrolution/compatibility_check_spec.rb
@@ -1,12 +1,12 @@
 describe Avrolution::CompatibilityCheck, :fakefs do
-  let(:schema_registry) { instance_double(AvroTurf::ConfluentSchemaRegistry) }
+  let(:schema_registry) { instance_double(AvroSchemaRegistry::Client) }
   let(:app_schema_path) { File.join(Avrolution.root, 'avro/schema') }
   let(:logger) { instance_double(Logger, info: nil) }
 
   before do
     allow(ENV).to receive(:[]).and_call_original
     allow(ENV).to receive(:[]).with('COMPATIBILITY_REGISTRY_URL').and_return('registry_url')
-    allow(AvroTurf::ConfluentSchemaRegistry).to receive(:new).and_return(schema_registry)
+    allow(AvroSchemaRegistry::Client).to receive(:new).and_return(schema_registry)
     FileUtils.mkdir_p(app_schema_path)
     # Diffy uses the tmp directory
     FileUtils.mkdir('/tmp')
@@ -82,7 +82,7 @@ describe Avrolution::CompatibilityCheck, :fakefs do
       context "when the schema is compatible using the defined compatibility break" do
         before do
           allow(schema_registry).to receive(:compatible?)
-                                      .with('com.salsify.app', Avro::Schema, 'latest', with_compatibility: with_compatibility).and_return(true)
+            .with('com.salsify.app', Avro::Schema, 'latest', with_compatibility: with_compatibility).and_return(true)
         end
 
         it "returns success" do
@@ -95,7 +95,7 @@ describe Avrolution::CompatibilityCheck, :fakefs do
 
         before do
           allow(schema_registry).to receive(:compatible?)
-                                      .with('com.salsify.app', Avro::Schema, 'latest', with_compatibility: with_compatibility).and_return(false)
+            .with('com.salsify.app', Avro::Schema, 'latest', with_compatibility: with_compatibility).and_return(false)
         end
 
         it_behaves_like "an incompatible schema"