+++ lib/avrolution/configuration.rb
@@ -1,6 +1,7 @@
 module Avrolution
 
   COMPATIBILITY_SCHEMA_REGISTRY_URL = 'COMPATIBILITY_SCHEMA_REGISTRY_URL'.freeze
+  DEPLOYMENT_SCHEMA_REGISTRY_URL = 'DEPLOYMENT_SCHEMA_REGISTRY_URL'.freeze
 
   class << self
     # Root directory to search for schemas, and default location for
@@ -14,6 +15,10 @@ module Avrolution
     # The URL (including any Basic Auth) for the schema registry to use for
     # compatibility checks
     attr_writer :compatibility_schema_registry_url
+
+    # The URL (including any Basic Auth) for the schema registry to use for
+    # deployment
+    attr_writer :deployment_schema_registry_url
 
     attr_accessor :logger
   end
@@ -29,9 +34,14 @@ module Avrolution
   end
 
   def self.compatibility_schema_registry_url
-    @compatibility_schema_registry_url ||= begin
-      raise 'compatibility_schema_registry_url must be set' unless ENV[COMPATIBILITY_SCHEMA_REGISTRY_URL]
-      ENV[COMPATIBILITY_SCHEMA_REGISTRY_URL]
+    @compatibility_schema_registry_url ||= ENV.fetch(COMPATIBILITY_SCHEMA_REGISTRY_URL) do
+      raise 'compatibility_schema_registry_url must be set'
+    end
+  end
+
+  def self.deployment_schema_registry_url
+    @deployment_schema_registry_url ||= ENV.fetch(DEPLOYMENT_SCHEMA_REGISTRY_URL) do
+      raise 'deployment_schema_registry_url must be set'
     end
   end